Add documentation and rename variable and config option
If the project is compiled in a different system than the one running vim,
then the path of the source file used when Jumping to Definitions (C-])
will be the one from the system where the code was compiled, instead of
the local path where vim is running.

This is usually the case when you are running your project inside a Docker
container and mounting the files inside the container or when you are
mounting a remote filesystem locally with FUSE

setting g:alchemist_compile_basepath to the base dir where your project was
compiled will allow us to translate it correctly to a relative path in the
local file system.

    let g:alchemist_compile_basepath = '/app/'